update flow backup wallet
